+++ CHANGELOG.md
@@ -9,8 +9,24 @@ release date when you use `npm version` (see `README.md`).
 
 ### Added
 
+- Main tab is now completely renewed (based on what was the mobile tab).
+  Instances and messages are integrated into this.
+
 - Added Content Security Policy for instances - the same one as DeltaChat
   Desktop.
+
+### Fixed
+
+- Reload button is now disabled when instance is not started.
+
+### Changed
+
+- Refactored implementation of mobile tab and cleaned it up.
+
+## Removed
+
+- Remove `--open` automatic open option as it makes little sense with the
+  new UI.
 
 ## [0.13.1][] - 2022-06-30