+++ CHANGELOG.md
@@ -7,6 +7,8 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+## [0.14.0][] - 2022-06-30
+
 ### Added
 
 - Main tab is now completely renewed (based on what was the mobile tab).
@@ -317,3 +319,7 @@ release date when you use `npm version` (see `README.md`).
 [0.13.0]: https://github.com/webxdc/webxdc-dev/tree/v0.13.0
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.13.1...HEAD
 [0.13.1]: https://github.com/webxdc/webxdc-dev/tree/v0.13.1
+
+
+[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/webxdc/webxdc-dev/tree/v0.14.0